+++ src/qnethernet/QNEthernetClient.h
@@ -200,6 +200,11 @@ class EthernetClient : public internal::ClientEx,
   // If this returns zero and there was an error then errno will be set.
   size_t write(const uint8_t* buf, size_t size) final;
 
+  // A convenience function that can write any type of data.
+  size_t write(const void* const buf, const size_t size) {
+    return write(static_cast<const uint8_t*>(buf), size);
+  }
+
   int availableForWrite() final;
   void flush() final;